+++ actionmailbox/lib/action_mailbox/gem_version.rb
@@ -9,8 +9,8 @@ module ActionMailbox
   module VERSION
     MAJOR = 6
     MINOR = 1
-    TINY  = 4
-    PRE   = "7"
+    TINY  = 5
+    PRE   = nil
 
     STRING = [MAJOR, MINOR, TINY, PRE].compact.join(".")
   end